requirements.txt
streamlit
openai
requests
pillow
pytesseract
beautifulsoup4
